Fixed infinite recursion in parameter creation

+++ nestkernel/nest.h
@@ -230,7 +230,7 @@ dictionary get_model_defaults( const std::string& model_name );
 // TODO-PYNEST-NG: static functions?
 ParameterPTR create_parameter( const boost::any& );
 ParameterPTR create_parameter( const double );
-ParameterPTR create_parameter( const int );
+ParameterPTR create_parameter( const long );
 ParameterPTR create_parameter( const dictionary& param_dict );
 ParameterPTR create_parameter( const std::string& name, const dictionary& d );
